2022 08 21

+++ c++/e399.cpp
@@ -3,7 +3,7 @@ using namespace std;
 int main()  {
     int n;
     cin >> n;
-    int a[n],b[n+1]={0};
+    long long a[n],b[n+1]={0};
     for(int i = 0; i < n; i++) {
         cin >> a[i];
         b[i+1]=b[i]+a[i];
